Do not put the destination ID of a packet in the packet itself
The destination ID of a packet should not be in the packet itself (the
payload), but should instead be provided separately. This patch moves
the destination ID out of the packet and instead allows a destination ID
to be provided to the `enqueue` method. Thismakes more sense and is also
more readable. Additionally it prevents some unnecessary data transfer
and destination ID handling.