Make use of the new custom XBee packet logic
Replace instances where we used regular XBee packets for custom messages
with custom XBee packets. Additionally we refactor the custom packet
handling and implement a `get_all` method for packets such that we can
print their contents directly. Previously we relied on serialization,
but for custom packets that would yield a byte-encoded string, which is
not really helpful for the prints.