MockVehicle: Use droneapi Locations
We wanted to keep this mock vehicle library completely separate from
droneapi, so that tests worked without the latter library and we ensure
that we do not secretly use anything from it. However, Location is
a very simple class in droneapi, and the tests would fail if mixed types
are given to the equality function. Thus use the class just so that all
the objects are the same type.