Mission_Pathfind: Avoid array index out of bounds
When we reach the final point, we should stop trying later points. The
step function would not start again in this case, but when it is already
within the function it might go out of bounds at the final point. Make
sure this does not happen.

We can later implement a return to launch here, but this is just to fix
it from raising an error.