Geometry.get_point_edges: Handle empty point lists
Make sure that an empty list does not result in an index error but
actually does what we expect, namely return an empty list of edges.

This fixes the memory map plot display for missions that do not have
waypoints, and is potentially helpful for badly formed VRML files that
have objects without any polygons -- those are ignored rather than raise
an error with this change.