Geometry: Fix home location for older versions of dronekit
The home location property is not set on the stable version, but the
documentation about it already appeared on the dronekit website.
Additionally, the home location is not set until after the commands have
been downloaded, so setting it in the Environment is too early.

Move the setting of the home location to the Geometry to Mission, where
we can retrieve the waypoint Location and pass it on to Geometry.