Ensure Location objects have correct relativeness
- When creating a projected location in Distance_Sensor_Simulator, apply
  relativeness of the original location.
- Equalize relativeness in Geometry 2D point in polygon check.

This is mostly because the Rover gives non-relative global coordinates
more often than a Copter does, possibly because it does not have an
altimeter et cetera. Eventually the problem lies in that the sensor is
simulated to be at an equal height as the lower edges of the polygons,
which makes them usually invisible. That needs a different solution.