Mission_Pathfind: Decrease waypoint list size
Make the reconstructed path not contain every single point from the
memory map that we should follow, but instead skip points in the case
that the trend is the same as before. This helps reduce the size of the
list which for some reason decreases CPU time in searching in the list,
and also makes the mission faster in the sense that it does not have to
reach each waypoint before issuing a command to go to the next point.

This is increasingly important for the memory map that can have an
increased resolution. Otherwise we would issue a waypoint for every
index which could become too close to each other.

This does make the path less precise and returns problems with the
vehicle passing through objects, but that can be amended by decreasing
the speed (which was artificially decreased due to the waypoint checks).