Environment: Check whether we flew through objects
Add a callback to MockVehicle when the location changes so that the
environment can check whether there was an object in between the
original location and the new one. Add a guard for recursive calls to
the callback. Move a bunch of methods from Distance_Sensor_Simulator to
geometry now that they are generalized enough. This could be helpful to
detect simulation problems.

This is however problematic since checking the objects takes time, which
means the location should already change once it is set. Thus this
actually encourages the vehicle to fly through objects because its
status not checked often enough.