mission_basic: Skip traceback on RuntimeError
A runtime error is generated by our own code as a means of mentioning
that we are at an unsafe position or problems with the memory map due to
which no path can be found. These should be handled the same as actual
errors such as memory map out of bounds as in stop and return to launch,
however we do not need to see where the error was generated since we
already know the reason why. This cleans up the output a bit.