MockVehicle: Fix attitude after takeoff
The simulated vehicle changed its attitude to a bearing of 90 degrees
during takeoff, which is unwanted since the actual vehicle does not do
that. This is caused by the fact that Geometry.get_distance_meters takes
altitude differences into account now, while set_target_location did not
expect it to do so. Instead just compare the latitude and longitude
components.