MockVehicle: Change exception handling in location check again
Instead of gracefully unsetting the location callback, just set it to
its prior callback so that recursion detection does not cascade. This
now means that the mission script will cause two exceptions for flying
into an object (if enabled), one for the actual attempt and one for the
return to launch since it still flew inside the object. We can handle
this wave of exceptions better but it is out of the scope of this
feature.