Geometry: Clean up plane intersection some more
The splitting of methods for plane distance and plane intersections was
useful, but actually we are always interested in polygon intersection,
so perform the projected point in polygon check as well. This reduces
the need of returning the plane coordinates, cleans up plane distance
more, and fixes the tests for flying into an object in the environment
so that it does not detect them too much.