Viewer: Allow toggling flying through objects
Make it possible to use the F key to make the polygons physical or to
allow flying through them. This is useful in testing whether the
location checks all work, and enables a more realistic simulation within
the interactive viewer. Also cleans up some weirdness about update
returning a Location object, which was too late for us to reset the
update when we are about to fly into an object. We can now do so.