Monitor: Be less strict about leaving memory map
We rather do not want to leave the mission search space, but it is not
helpful to immediately end the mission if it happens. It might indicate
some sensor problems, but the mission itself could still be finished. If
need be, the ground station should decide to stop the vehicle at some
point. Instead of stopping the mission with a key error, log it so we
can check it later on.