Mission: Skip adding takeoff command for rover
Do not add a takeoff command to the autopilot. The autopilot may or may
not attempt to adhere it, and fail to start the mission if it does so.
Instead, set the altitude to zero to prevent other problems with
following waypoints. We could also disable all altitude components when
running as a rover, but this should do for simple missions for now.