Mission: Allow disabling GPS fix
Add a setting to determine whether we wait for a GPS fix before starting
the mission. We can turn this off when the rover is inside a building
with too many concrete floors above it for example. Not having a GPS fix
is quite bad since the current code depends on having Location objects,
but the autopilot should still be able to do something. Whether it is
sensible is another concern. We can later use local positioning for our
missions, but this requires some rewrites.