Mission: Better command checks
We previously send the commands during setup, which is too early since
it might actually be cleared or redownloaded in between. We can hold it
off until arming. However, the physical autopilot is too slow to receive
the commands list, which makes it only start sending commands when we
want to set AUTO mode. The autopilot then refuses to change mode since
there are no commands. Make use of the delay in display to ensure its
success for now. In dronekit 2 we wait for the upload with a function.

Also clean up the home location retrieval since it might change between
setup, clear, redownload, adding our own mission and arming. We want the
latest home location since it might be fixed by GPS, thus update it in
the Geometry when we check the mission update.

Also add more prints for the locations which is helpful for debugging.