Reconstruction_Planning: Initial framework setup
We want to plan our trajectories in our missions by using the
reconstruction problem as a goal. We can optimize the coverage of the
measurements and the duration of the mission using optimization
algorithms. In this case, we use multiobjective programming with
evolutionary algorithms to find an approximation of the optimal
solutions. This uses objective functions and constraints to optimize the
solution variables (which in our case will be the parameters of lines
that we want to measure at). The variables are mutated using an
evolutionary algorithm that discards solutions that are worse in all
objectives than others to make room for new mutated possible solutions.

The current structure is mostly an import of an old project of me, minus
the actual problem definitions. The code now has documentation comments
rather than just implementation decisions. It has also been split into
a script (that does not yet do anything) and the class framework.