Snap to boundary: properly take care of horizontal and vertical lines
Previously, horizontal and vertical lines could cause a division by zero
in the code. This patch changes the `is_intersecting` method to first
determine the type of line we are working with (horizontal, vertical or
other) and then base further calculations on this information. As an
additional bonus, this makes sure that we only check the necessary
boundaries and not more than that.

The existing unit tests still pass and two more tests have been added to
make sure that horizontal and vertical lines are also handled properly.