Snap to boundary: do not snap points that are already on a boundary
This patch makes sure that we do not try to snap points that are already
on a boundary. As Leon showed, doing that might cause issues especially
with points that are on a horizontal boundary because of the
`math.atan2` and `math.tan` involved. Since there is no need to even try
to snap them, it is not only safer but also faster to ignore them
immediately.