Implement a dump reader
This patch allows us to read dump files for usage in the reconstruction
and visualization phases. These dump files contain the data we need in a
much easier format than the previous datasets we used, because this
format corresponds to the XBee packets that the ground station
currently receives. The reader is therefore also easier and makes it
possible to obtain valid XBee packets from the dump file, just as if
they came in via a stream. Therefore this work makes it easier to use
streaming data later on because the dump files are actually interpreted
as a stream.

We remove the old reader and associated files and convert our runner to
simply output the packets for now.