Planning: Variable to snap to cardinal direction
Add another variable that causes the corresponding line to become
a horizontal or vertical line if it is close to 0 or 90 degrees,
respectively. Previously, such straight lines had a probability close to
zero to even appear, since it is infinitely more likely that the angle
variable value is something other than exactly 0.00... or pi/2 radians.
This variable is a boolean that, when turned on, snaps angles that are
pi/8 radians (22.5 degrees) away from a cardinal direction to that
direction. This increases the search space which is actually wanted, so
that we can get more results with such straight lines.