Weight matrix: refactor grid creation
There is no need to create the grid only for the space covered by the
sensors. Doing so has the disadvantage that we cannot cache the grid and
the associated distance measurements because when a new signal strength
measurement comes in that extends the space, everything has to be
recalculated.

Instead we now create the grid for the space covered by the network. We
also use the centers of the pixels now, rather than the edges, for
calculating the distances. This makes sure that the distance between
grid points is one all the time. Previously we would have that a linear
space from 0 to 3 with three items became [0, 1.5, 3], whereas we now
get the evenly spaced [0.5, 1.5, 2.5].