Rover: Remove mission scripts
These scripts are no longer used when starting the mission using
Dronekit 2, and were already not mentioned in the README. The
ARMING_CHECK parameter setting was moved to the Mission python code
itself (although it seems not all ArduPilot have this parameter anyway).
The default copter- or rover-specific parameter files should simply be
loaded once when assembling the vehicle, not every time the script is
started. This used to cause any changes to parameters to be overidden.

Therefore these scripts have no real or documentation purpose anymore.