Vehicle: Create interface for Vehicle controllers
Determine the required methods and properties based on the Mission
class. Probably more methods will be necessary based on e.g. the Setup
in mission_basic and Envrionment, but this should at least encompass the
mission interface. The Vehicle is not yet used anywhere, but it will be
used as base class for the dronekit API vehicle, MockVehicle and other
controllers we want to support.