Robot_Vehile: Split out direct connection code
Move the code that uses the motor speed servos and direction pins
directly to Robot_Vehicle_Raspberry_Pi. It also uses the
Line_Follower_Raspberry_Pi of which we can only use in this way as well.
This allows us to add other versions of robot vehicle control, such as
serial or PWM connections.