Arduino: Add simple serial connection program
Add a directory for Arduino scripts, most notably scripts that allow us
to interface with a Zumo robot via a serial connection to the Arduino.
This allows us to make use of the Zumo library as well as our own python
components that handle the heavy lifting on the Raspberry Pi.

The particular script here starts a serial connection and send sensor
measurements over it. It also reads from the serial interface if
possible. For now all data read is echoed back. This can allow a vehicle
controller to change motor speeds, without any disruption for the line
follower module, for which the data travels in the opposite direction.

This script is based on and requires the Zumo Library available at
<https://github.com/pololu/zumo-shield>.