Robot_Vehicle: Change speed property meaning
Instead of providing the direct speed value according to servo PWMs or
other ways of acquiring this measurement, use the speed getter/setter
for the movement speed during the normal state of moving. This means
that we can actually ensure that the movement speed is retaining during
the different states that we are in, and not reset to some weird value.
This should also make GUIDED missions work since they receive this speed
but first need to stand still at an intersection.