Robot_Vehicle_Arduino: Initial version
The Arduino version of the robot vehicle communicates with an Arduino in
order to set the speeds of the motors. For now we use the same serial
object as the line follower. Even using multiple instances for the same
serial device should not cause problems since the robot vehicle only
writes data while the line follower only reads data.