XBee: Compatibility with pySerial 3.0
Make the XBee component compatible with the newest version of pySerial.
Previously, version 3.0 would give an IOError for some internal control
flag setting. The fix appears to be to enable the RTS and DTR control
lines at initialization. This makes the tests pass with Serial 3.0
installed. Also tested with the physical XBees.

Updated the README and travis dependencies to match.