Mission: Add initial robot grid fan cycle mission
Mission_Cycle lets two robot vehicles move around in such a way that the
XBee measurements create fan beams and straight lines. This is a good
first approach to using the grid layout for making these measurements.

This requires some more synchronization with the XBee status packets
though, so that the robots wait until the measurement is done before
moving to the next grid point.