Line_Follower_Arduino: Compatibility with pySerial 3.0
The Serial connection in the line follower also needs the `rtscts` and
`dsrdtr` parameters per #51. However, this is only the case for serial
devices that are not the actual serial USB, such as virtual devices and
perhaps other devices than the one that we use. Use the parameters to
make sure the connection always works.