Tests: Split line follower tests into own methods
This has the advantage that the initialization/reset of the line
follower object and the magic mock callback is done in one setUp method,
rather than duplicated. This makes it easier to add more parameters
later on. Additionally, since they are separate tests, they are also
seen as their own tests and count toward the total.