Test: Return test result status in test.sh
Since we started using test.sh, Travis has not been very good at
determining whether our tests pass or fail. This is because the status
that Travis reports depends only on the return value of the command that
it runs, not any error output.

Fix this problem by keeping the status value and return it at the end of
the test script.