Tests: Reset settings in tests
This makes sure that changes through settings via Arguments or
Settings.set do not bleed into other tests. This is caused by the
internal static cache that Settings keeps. The settings and arguments
test cases took account for this, but others did not. Thus, the mangling
used to happen all the time with nobody noticing because tests usually
focus on one component.

Now, with tests that require the use of multiple components such as the
Mission_Cycle test, changes to settings may overlap. This could cause
a test to run with a different vehicle than the default Mock_Vehicle,
causing unexpected results.

This leads to intermittent test failures (which were masked by the test
runner not returning a correct status value). Add a SettingsTestCase
which test cases can inherit its tearDown method from. Add some super
calls where necessary or where it is future-proof.