Robot_Vehicle_Arduino: Remove DTR changes
The idea of sending pulses on the DTR line of the serial connection is
interesting, since it allows us to reset the Arduino based on the value
on this line. However, we have not yet used this so far, since resetting
is not immediately required to work with the Arduino. It should still
function normally when only the Raspberry Pi resets its connection.

Additionally, the use of DTR causes problems for devices that do not
have it. This includes virtual devices such as terminal forwarding or
the python pty library. This means tests fail only due to not being able
to send the message, even though we do not use it physically right now.

Remove the uses of the DTR setter. In case we want it back later, it
would need to be changed to ensure the signal is appropriate for
resetting the Arduino, which requires more physical testing.