Vehicle: Add is_(current_)location_valid methods
Add two methods for checking whether a location is populated with
a somewhat correct location. Vehicles can reimplement these to support
other location object types or perform other validations of the location
properties. The Robot_Vehicle considers the location to be invalid while
moving, but this can be extended with other checks later on.

Also cleaned up some documentation and method body style.