XBee: Add location valid field and callback
Add another argument to the XBee_Sensor to determine whether the current
location is valid or not. This can then be passed along with the packets
so that another vehicle knows whether the measurement is valid, and
likewise for the ground station. This allows us to synchronize the
vehicle measurement locations and to discard any measurements that were
made while vehicles are moving.

This only adds the fields and callbacks, as well as some small cleanup.
The actual synchronization of the mission will follow in another commit.