Environment: Track valid measurements, add docs
If both the current vehicle's location and the other XBee's location are
valid, then the measurement is complete and the mission should be able
to detect this. Therefore, track this situation and add a method that
exposes the member variable.

Additionally, add a bunch of documentation for methods detailing their
arguments and return values, and which situations they depend on, and
add safety checks to add_packet_action.