Mission_Cycle: Wait for valid measurements
Only start moving to the next waypoint if both the current vehicle's
location is valid and the same for the other vehicle, according to the
received packet.

For the Robot_Vehicle, the location is now valid if we are not moving
and the current waypoint is the current location. This means that valid
locations are only received in AUTO or GUIDED missions that use, e.g,
simple_goto, and not manual missions such as Mission_Infrared. We could
enable them for missions without waypoints later on, but synchronization
becomes more difficult in that case since we need to be at a specific
intersection, not just an intermediate one.