Arduino: Initial zumo grid code
Based on the Zumo maze solver code, and written by Folkert Bleichrodt,
with some small cleanup.

This does not interface with either versions of the Robot_Vehicle, since
it uses the line follower on its own. It will need some changes to the
serial interface before we can make use of it from the Python code.