Control_Panel: Clean up menu bar code
Move the menu bar code to the controller where it is more suited, since
the view is being cleared. Remove duplication of adding actions to the
View menu, and add a check mark to the current view in the menu. Do not
switch views if the current view is already this view, which makes the
view keep its state in this case at least.

Also document how we can add additional menus and how we should reset
the state when the view is cleared if we do so.