Prefix control panel settings with the view name
This makes it more clear to see which settings belong to which view.

Additionally, this patch improves the loading view code by not
constantly getting the settings and the loop delay, but only doing it
once. This makes the view react slightly faster to XBee device
insertions. We also use `sys.exit(1)` instead of `self.window.close()`
because the latter would not actually exit the entire application when a
runtime error occurs. I noticed this when I used an incorrect setting
name. It would show the traceback, but after closing the message box the
loading view would still be visible.