Control panel: implement sensor ID check
This patch makes the loading view check if a ground station XBee has in
fact been inserted (which has ID 0). Any other XBees will result in a
message box being shown.

For this we had to refactor the physical XBee sensor component a bit. We
split the setup phase into two phases: identification (for fetching
the node identifier and the address) and joining (entering the network). This makes it possible for the loading view to only request the sensor's credentials without already letting the device attempt to join the network.

Finally, the first view is now the waypoints view. This makes more sense
for the mission use case, where we first want to enter waypoints and
    then consider the reconstrution.