XBee TDMA scheduler: only accept future timestamps
During initialization, some XBees are activated before others. In that
case they already start sending packets, but the other sensors do not
receive them yet. When the other sensors are also activated, these
packets are received, but at that time the timestamps in those packets
are old. Before this patch we would use those old timestamps for
synchronization, causing the network to be flooded with packets because
of a wrong scheduling. This patch prevents that by only synchronizing
with future timestamps.