XBee TDMA scheduler: correct and improve the unit tests
Some time ago we changed the number of sensors in the settings from
eight to two. However, the unit tests still expected eight sensors for
the synchronization tests. This patch makes sure that the unit tests use
eight sensors such that they pass again after the last commit.

Furthermore we can use `assertEqual` because the synchronization tests
are not timing dependent.