Arduino: Convert zumo_grid to SoftwareSerial
For some reason the 0/1 pins do not always work correctly on some
Arduinos, so make it possible to use other pins for the serial
connection. Use defaults that do not clash with default uses of the Zumo
shield (6 is only used for Leonardo/A-Star and depends on the buzzer
jumper, A1 is only used if the voltage level jumper is soldered).

Add functions that make life easier with the simplistic interface of
SoftwareSerial, such as ensuring data is available before reading it,
ignoring input until EOL, reading strings and parsing integers with
a read-until-newline safety check.