Implement node discovery for vehicle XBees
This completes the devices view. When we first open the view, the ground
station device data is shown. The refresh button yields the information
for the other vehicle XBees using a node discovery process.

I have tested this by starting the control panel with only the ground
station inserted. After that, I inserted the two other XBee devices,
started a serial connection to them in a separate terminal (which is
required to activate them and let them join the network), and pressed
the refresh button. Both devices are found pretty quickly. It is also
nice to see that sometimes one device joins the network faster, in which
case we see the information for that device first and need another
refresh for the second device.