Control panel: fix a regression in the reconstruction view
Recent patches introduced a regression in the reconstruction view. We
would not be able to perform a second reconstruction because the view
was not cleared properly.

To resolve this, remove the hacky `clear`/`show` construction and
instead make the label and layout only once in the `show` method. This
not only fixes the regression, but is also more efficient as we do not
need to redo the label and layout constantly, only update it.