XBee sensor simulator: do not use 100% CPU when no data is available
When there is no data in the pipeline, we used to just continue the
loop, but by doing so we skipped the necessary loop delay a few lines
below it. This patch adds the loop delay in that path and thereby
resolves the 100% CPU usage. Note that 100% CPU was only achieved when
using the simulator component, not when using the physical XBee
component.

We did not notice this defect earlier because the runner would always
provide the simulated sensors with data, so we would never reach this
code path. For the control panel use case, however, they receive data
only sometimes, which made us notice this issue.