Arduino: Clean up reset/activate and comments
The DTR reset line is quite tricky to get a hold of, but most likely
this is how it should work. We must always start the Zumo before the Pi
to prevent a small amount of power being sent to the Arduino's reset
line and the brown-out detector kicking in.