XBee sensor simulator/physical: update runners and increase throughput
Recent optimizations allow us to send five packets per second instead of
four.

We update the runners to not send custom packets anymore. It is
problementic to time it properly because the physical XBee sensors take
a variable time to initialize. We test custom packets all the time using
the GUI, so there is no need for them in the runners.