Robot_Vehicle_Arduino_Full: Correctly handle serial exceptions
This version of the robot vehicle has a special serial loop thread.
Previously it was in the same thread as the state loop, and special
exception handling was added to prevent it from blocking.

However, this strategy was suboptimal and did not work in practice, so
it was moved to its own thread which is allowed to block. But then the
exception handling remained unchanged. This could cause the thread to
run longer than it needs to when the connection is closed but the thread
is not yet deactivated.

This edge case does not happen in runners, and tests do not use this
vehicle class directly. This brings it more in line with other thread
loops and can help writing later tests.