USBManagerTestCase: Clear all serial connections
In 3e9b4c29af0671a6d3e68dffc22892e371643b30, we close the "other" side
of the virtual port which is used in the Mission_Cycle test to check
whether the vehicle sends anything on the serial connection. However,
the "hub" side of the port was not closed.

One could argue that the tests should do this themselves, or the thread
manager closes the interfaces implicitly by stopping the Threadable
objects that use them. That is not good practice since the USB test
setup created them in the first place.

Note that this may cause problems for tests when they attempt to close
the interfaces themselves. They can be simplified since they no longer
need to do so in the tearDown method. The call to deactivate in the
Mission_Cycle test is one such example, which is even more dubious since
the thread manager handles that as well.