Settings: Step 1 of conversion to defaults
Add a new JSON file based on the old settings.json with details for all
settings such as help, explicit type and type-specific properties
(length and subtype for sequence types, min/max for numeric (sub)types,
and format/replacement table and required state for string types), and
list of allowed options or source keys for these options.

The types include bool, int, float, string, list and tuple (the latter
two comprising a new distinction) as well as two new string types known
as file and class. These will allow easier location of usable files and
selection of class options from a module dictionary, respectively. There
also exists a dict type, which will not be alterable from command line.

This is a first step to converting the settings into multiple files. The
defaults is augmented with settings file that override the defaults,
such as development, production or vehicle-specific settings, which can
in turn be overridden by command line arguments. The XBee specifications
are not in the defaults, since they are not intended to be alterable.
They will be moved to their own JSON file in a following commit.

The old settings file is kept in this commit to keep everything working,
and will be removed once the settings rewrite is in a later stage.