Settings: Use defaults, clear overrides (Step 3)
Convert the Settings class to use the defaults.json file to read in
defaults as well as the settings.json to fetch any overrides. The
settings are now cleared so everything is default, but we can survey
this later in case there are settings that have non-sane defaults.

Update the tests to cope with this. Specifically, allow them to pass
a defaults file and a settings file to Arguments and to Settings, and
use additional specialized test files for this. Add new tests for this
functionality as well as for some new methods of settings to ease
retrieving information about the Settings by Arguments, to be used soon.