Settings: Apply new definitions (Step 4)
Start making use of the information in defaults.json in Arguments and
Settings. Add help (or fall back to pretty option name if no "help"),
list of choices (from "options", an attribute pointed by "keys" or
module exportable attributes from "module"), the requested "length" and
"subtype" for list and tuple types, and the "replace" translation table
to the argument parsing. After parsing, ensure the type is correct
before passing it on to Settings.set, and handle any errors from there.

Settings.set provides the remaining constraint checks, namely "required"
(which means nonempty in this context, not necessarily passed in the
arguments), "min" and "max" for int/float types, and the "format" for
the file type, which depends on its exact workings based on "required".

Fix some lost/missing information in the defaults, and update the tests
so that they still work and test for new features. We now specify
exportable submodules/classes for certain modules, which only contains
usable classes. Notably, this causes the Vehicle and Robot_Vehicle tests
to not work with the Arguments overrides, since those are base classes.
However, using Settings.set here works since it does not care about the
allowable list of choices.