Control_Panel: Split up settings
Move view-related settings into their own component. This is consistent
with how we create other components. However, we still need to load
every component so that the argument parser is populated at the start.
This is done by managing the settings for the views in the Controller,
which are passed to the View when it is created, simplifying their
access to their own settings. This should also make it easier to reload
settings in the control panel when they are changed.

The reconstruction view was using both a setting from the control panel
and settings from the reconstruction component. Now, the control panel
reconstruction component has the reconstruction component as its parent,
inheriting all its settings and thus making them available in arguments.

Finally, use QApplication.arguments to fetch our arguments, and ignore
any Qt-specific arguments. Although there are a bunch of these at
http://doc.qt.io/qt-4.8/qapplication.html#QApplication they do not seem
to have an effect in the control panel, but it is best to not touch them
ourselves.