Control_Panel: Introduce Settings view
This commit adds a view to the control panel which displays a selectable
list of components and a settings pane for the selected component. All
types of settings are supported, each with their own validation schemes
and widget format. Only the file type does not yet have a specialized
form input. Even lists, tuples and dictionaries with subtype definitions
are supported, with options to add or remove items in them if the type
allows to.

We display most of the available information, such as a command link
button to the parent component, names, types, help descriptions, and so
on. The form has an aesthetic look, similar to other applications with
advanced settings windows. Validators and comboboxes ensure only correct
data is entered.

The form can be altered freely, but currently none of the changes are
collected nor saved to any location. This follows in a later commit.