Control_Panel_Settings_View: Add filter search
Add the bottom button stack with a filter input and a dummy Send button.
The filter causes the components list to be filtered to components that
have the search text in their name or in a key, value or description of
one of their settings. If possible, the best matched setting is scrolled
into view. Clearing the search (e.g. by using the widget's clear button)
brings all the components back.

This allows us to use the settings view without worrying that we would
be unable to find the setting we are looking for. It is also aesthetic.