Thread_Manager: Do not interrupt main thread from unregistered threads
When a ThreadtableTestCase is finished, we destroy the thread manager.
This deactivates all the threads and unregisters them. However, some
threads may be blocking/unresponsive for some time, causing them to
check that they are deactivated much later. Even worse, they might hit
an error, such as a closed connection, in between. As normal, the error
interrupts the main thread, which might be performing another test or
have exited already. This respectively cause a KeyboardInterrupt or
segmentation fault.

No longer interrupt the main thread for threads that are (already)
unregistered. Add tests for this, and add a test for the log method as
an extra.