Thread_Manager: Do not log destroy when not in exception context
To prevent spam of "None" errors in the logs, skip the log call in
destroy when there is no info about the exception. Since we use destroy
when ending programs normally as well, the logs become too cluttered
with non-errors due to this. The interrupt method will still
unconditionally log, since it has useful information about the source of
the call and is specifically targeted for exception handling.