Test runner: Display logs before clearing them
In the future, to be able to find out why tests fail intermittently, the
logs may provide useful details about exceptions that are handled by the
interpreter but still cause spurious errors. These should now be mostly
resolved by the Thread_Manager changes. The logs are now at least useful
to print, so this seems like a helpful idea.