Settings view: Small background color cleanup
Changed the method that adds or replaces the current background color so
that it can also remove the color; renamed it to set_background_color.
Also add an accessor for the current color. Use these in the file widget
to clear the color before opening the dialog, so that the padding added
by the button to prevent text from overflowing under it is not lost.

Also reset the color to the state before opening the dialog in case the
cancel button was pressed.