Default settings: resolve a name clash
The new `settings.json` file is just a list of overrides without
mentions of component names to which they belong. Both the XBee sensor
simulator and the XBee sensor physical use a `port` setting, so
overriding one of them in the settings view causes the other to be
broken. We resolve this by making the settings unique.