Settings view: Scroll to match when changing panel
When you filter on a setting such as "port", the first component view is
selected with the best match scrolled into view. If you then change to
another component view, this one is again at the top which is not very
nice. This already used to be fixed by scrolling to the match every time
a view changes, but it was somehow lost in a rewrite. Therefore,
reintroduce this scroll to match upon component change.