Settings view: Track default/value after remove/add from list
This makes it so that if an item is removed and then added, the old
value is still available. It also ensures that if we remove an item,
save it, reload everything and then readd the old, default value, the
setting is actually removed from the ground station overrides rather
than uselessly set to the default.

Also remove an unused import.