Loading view: Clean up progress display
After inserting the ground station XBee, the loading view hangs a long
time on "Activating ground station XBee and network...". Also, when an
XBee is already inserted, no window is shown until the devices view is
shown. Both of these can take an awkwardly long time.

As it appears, the setup method can take quite some time. Therefore,
turn the indeterminate progress bar (which at that point would have
frozen because we do not update the UI often enough) into one with
multiple steps, indicating we have progress. Make identification and
activation two steps, and correct labels/comments for that.

The window not appearing is caused by the insertion loop immediately
taking control of the event loop, thus never updating the UI. Instead,
make the first call to the loop also start from a (short enough) timer,
thus freeing the UI to draw its window.

Note that for simulated XBees, this causes a flash of a progressing bar,
and for physical XBees the activation step is so fast that it is hard to
read the label. This should not be that much of a concern since being
fast is better than appearing to be unresponsive.